+++ constants.js
@@ -1,0 +1,6 @@
+
+module.exports = {
+  WEBSITE_PATH_PREFIX: "CapoeiraSongbook",
+  DEPLOYED_SONG_DATA_FOLDER: "song-data",
+  SEARCH_RESULTS_OUTPUT_FOLDER: 'search-index'
+}
