+++ search.11tydata.js
@@ -1,0 +1,7 @@
+const constants = require('./constants.js')
+
+module.exports = {
+  eleventyComputed: {
+    SEARCH_RESULTS_OUTPUT_URL_PATH: '/' + constants.SEARCH_RESULTS_OUTPUT_FOLDER
+  }
+};
