+++ search.liquid
@@ -1,0 +1,22 @@
+---
+layout: page-menu-banner.liquid
+title: "Capoeira Songbook|Search"
+---
+<h2>
+    Search
+</h2>
+
+<div class="hulipaa_searchbar"></div>
+
+<script type="module" src="https://cdn.jsdelivr.net/npm/hulipaa-ui@1.0.2"></script>
+
+<script type="module">
+    import { parseSongFileContent } from "{{ '/frontend/parse-song-content.mjs' | url }}";
+
+    window.addEventListener('DOMContentLoaded',() => {
+        Hulipaa({
+            parsePage: parseSongFileContent,
+            resultsPath: '{{ SEARCH_RESULTS_OUTPUT_URL_PATH | url }}'
+        })
+    });
+</script>
